+++ packages/core/src/models/module.ts
@@ -9,16 +9,16 @@ import {
   addImportDeclaration,
   updateImportDeclaration,
 } from '../helpers';
-import { TangoFile } from './file';
+import { AbstractFile } from './file';
 import { IFileConfig, IImportSpecifierData, ImportDeclarationDataType } from '../types';
-import { IWorkspace } from './interfaces';
+import { AbstractWorkspace } from './abstract-workspace';
 
 /**
- * JS 模块实现规范
+ * JS 文件抽象基类
  * - ast 操纵类方法，统一返回 this，支持外层链式调用
  * - observable state 统一用 _foo 格式，并提供 getter 方法
  */
-export class TangoModule extends TangoFile {
+export abstract class AbstractJsFile extends AbstractFile {
   ast: t.File;
 
   /**
@@ -31,7 +31,7 @@ export class TangoModule extends TangoFile {
    */
   importList: ImportDeclarationDataType;
 
-  constructor(workspace: IWorkspace, props: IFileConfig, isSyncCode = true) {
+  constructor(workspace: AbstractWorkspace, props: IFileConfig, isSyncCode = true) {
     super(workspace, props, isSyncCode);
   }
 
@@ -141,8 +141,8 @@ export class TangoModule extends TangoFile {
 /**
  * 普通 JS 文件
  */
-export class TangoJsModule extends TangoModule {
-  constructor(workspace: IWorkspace, props: IFileConfig) {
+export class JsFile extends AbstractJsFile {
+  constructor(workspace: AbstractWorkspace, props: IFileConfig) {
     super(workspace, props, false);
     this.update(props.code, true, false);
