+++ packages/core/src/models/node.ts
@@ -1,18 +1,17 @@
-import { JSXElement, SourceLocation } from '@babel/types';
+import { JSXElement } from '@babel/types';
 import { Dict } from '@music163/tango-helpers';
 import { cloneJSXElement, getJSXElementAttributes } from '../helpers';
-import { ITangoViewNodeData } from '../types';
 import { TangoViewModule } from './view-module';
-import { IViewNode } from './interfaces';
+import { AbstractFile } from './file';
 
-type TangoNodeConstructorPropsType = ITangoViewNodeData & {
-  file: TangoViewModule;
-};
+interface IViewNodeInitConfig<RawNodeType = unknown, ViewFileType = AbstractFile> {
+  id: string;
+  component: string;
+  rawNode: RawNodeType;
+  file: ViewFileType;
+}
 
-/**
- * 视图节点类
- */
-export class TangoNode implements IViewNode {
+export abstract class AbstractViewNode<RawNodeType = unknown, ViewFileType = AbstractFile> {
   /**
    * 节点 ID
    */
@@ -23,24 +22,55 @@ export class TangoNode implements IViewNode {
    */
   readonly component: string;
 
-  readonly rawNode: JSXElement;
+  readonly rawNode: RawNodeType;
 
   /**
    * 节点所属的文件对象
    */
-  file: TangoViewModule;
+  file: ViewFileType;
 
+  /**
+   * 节点所属的文件对象
+   */
   props: Record<string, any>;
 
-  get loc(): SourceLocation {
-    return this.rawNode?.loc;
-  }
+  /**
+   * 节点的位置信息
+   */
+  abstract get loc(): unknown;
 
-  constructor(props: TangoNodeConstructorPropsType) {
-    this.file = props.file;
+  constructor(props: IViewNodeInitConfig<RawNodeType, ViewFileType>) {
     this.id = props.id;
     this.component = props.component;
     this.rawNode = props.rawNode;
+    this.file = null;
+  }
+
+  /**
+   * 销毁当前节点，清空文件和节点的关联关系
+   */
+  destroy() {
+    this.file = undefined;
+  }
+
+  /**
+   * 返回克隆后的 ast 节点
+   * @param overrideProps 额外设置给克隆节点的属性
+   * @returns 返回克隆的原始节点
+   */
+  abstract cloneRawNode(overrideProps?: Dict): RawNodeType;
+}
+
+/**
+ * 视图节点类
+ */
+export class TangoNode extends AbstractViewNode<JSXElement, TangoViewModule> {
+  get loc() {
+    return this.rawNode?.loc;
+  }
+
+  constructor(props: IViewNodeInitConfig<JSXElement, TangoViewModule>) {
+    super(props);
     this.props = getJSXElementAttributes(cloneJSXElement(props.rawNode));
   }
 
@@ -51,12 +81,5 @@ export class TangoNode implements IViewNode {
    */
   cloneRawNode(overrideProps?: Dict) {
     return cloneJSXElement(this.rawNode, overrideProps);
-  }
-
-  /**
-   * 清空节点的指向，交给 GC 去回收
-   */
-  destroy() {
-    this.file = null;
   }
 }
